Fix a string-related TypeError in Python 2
Adding `from __future__ import unicode_literals` to `paginator_test.py`
introduced a `TypeError: __name__ must be set to a string object` from
`@functools.wraps`.

`functools.wraps(wrapped)` expects `wrapped.__name__` to be a unicode
string in Python 3, but to be a byte string in Python 2, otherwise it
raises `TypeError`.

`str` is the unicode type in Python 3 but is the byte string type in
Python 2, so pass `__name__` through `str` to turn it into a unicode
string in Python 3 (i.e. do nothing - it's already a unicode string in
Python 3) but turn it into a byte string in Python 2. This keeps
`functools` happy in both Python 2 and 3.

Once we no longer support Python 2 we can remove this.