Add logged in user to join template test
The test test_renders_join_template_when_no_read_permission was
previously incorrect in that it did not set a logged in user.
This happended to work before because the if statements were run
in a different order however, now the not logged in user check
happens before the join permissions check. Without a logged
in user set for the test, the code exits early due to the not
logged in user check. This has been fixed by adding a user to the
test's request. The other join tests fake permission functions
were also simplified due to the reordering, and again a user was
added so that the code doesn't return early. There are 3 cases
here that need to be tested: the user is not logged in, the user
is logged in and doesn't have read permission but does have join
permission, and the user is logged in and doesn't have read or
join permission.